메인레이아웃에 통합
