--- mzpick-front/src/layouts/SideBarLayout/style.css
@@ -1,83 +1,0 @@
-.App {
-
-}
-
-#main {
-    display: flex;
-    justify-content: flex-end;
-
-}
-
-#main .side-bar {
-    display: flex;
-    flex-direction: column;
-    background-color: var(--primary-color);
-}
-
-#main .side-bar .closeIcon {
-    display: flex;
-    margin: 10px 24px 30px 236px;
-    width: 40px;
-    height: 40px;
-
-    background-image: url('/public/image/Close\ Icon.png');
-    background-size: contain;
-}
-
-#main .side-bar .Category {
-    display: flex;
-    width: 252px;
-    height: 50px;
-    font-size: 17px;
-    color: var(--primary-font-color);
-    margin: 0px 24px;
-    align-items: center;
-    border-bottom: 1px solid var(--primary-font-color);
-    cursor: pointer;
-}
-
-#main .side-bar .darkLine {
-    border-bottom: 1px solid rgba(201, 224, 253, 0.3);
-}
-
-.signin-signup {
-    display: flex;
-    gap: 5px;
-    height: 300px;
-    justify-content: center;
-}
-
-.signin-signup .login-button,
-.signin-signup .slice-line,
-.signin-signup .signup-button {
-    display: flex;
-    height: 50px;
-    font-size: 16px;
-    align-items: center;
-    color: var(--primary-font-color);
-    cursor: pointer;
-}
-
-.signin-signup .slice-line{
-    cursor: default;
-}
-
-#main .side-bar .Travel-detail {
-    display: flex;
-    flex-direction: column;
-    height: 125px;
-    margin: 5px 152px 5px 24px;
-}
-#main .side-bar .Travel-detail .Travel-detail-text-box{
-    display: flex;
-    flex-direction: row;
-}
-
-#main .side-bar .Travel-detail .Travel-detail-text-box .icon{
-    color: rgb(76, 148, 211);
-}
-
-#main .side-bar .Travel-detail .Travel-detail-text-box .Travel-detail-text{
-    color: rgb(76, 148, 211);
-}
-
